Run db_stress for final time to ensure un-interrupted validation
In Blackbox tests db_stress command always run with timeout.
Timeout can happen during validation, leaving some of the keys
not checked. Since key validation is done in order, it is quite
likely that keys are towards to the end of the set are never validated.
This PR adds a final execution, without timeout, to ensure validation
is executed for all keys, at least once.